Issue #225 (Bug) is fixed.  Changes to be committed: 	modified:   persepolis/gui/addlink_ui.py 	modified:   persepolis/gui/mainwindow_ui.py 	modified:   persepolis/gui/progress_ui.py 	modified:   persepolis/scripts/download.py 	modified:   persepolis/scripts/mainwindow.py 	modified:   persepolis/scripts/progress.py 	modified:   persepolis/scripts/properties.py

+++ persepolis/scripts/mainwindow.py
@@ -3936,7 +3936,7 @@ class MainWindow(MainWindow_Ui):
                 limit_unit = limit[-1]
 
             # limit_spinBox
-                self.limit_spinBox.setValue(int(limit_number))
+                self.limit_spinBox.setValue(float(limit_number))
 
             # limit_comboBox
                 if limit_unit == 'K':
